Add package gluon-mesh-vpn-wireguard (Part B)
The address of the vpn interface is calculated in the style of modified 
EUI-64, based on a virtual mac address. This virtual mac address 
consists of 0x00 as first byte and the other five bytes are taken from 
the first bytes of md5sum(base64 encoded public key).
    
The algorithm was taken by the ffmuc, with a slight difference. ffmuc 
calculated took the result of md5sum(base64 encoded public key + '\n') 
which was interpreted as accidential fault and therefore dropped.
    
Example:
- Public-Key: "gP3VJnTTvnQut+z4O+m0N9RgMyXbgyUbUkF3E3TKX2w="
- Address: "fe80::02ca:b8ff:fedc:2eb3"